tzif: posix TZ arithmetic

+++ src/zoned_datetime.rs
@@ -38,7 +38,7 @@ impl ZonedDateTime {
         zoneinfo: *mut PyObject,
     ) -> Option<ZonedDateTime> {
         let ordinal_secs = i64::from(date.ord()) * i64::from(S_PER_DAY)
-            + i64::from(time.total_seconds() - offset_secs);
+            + i64::from(time.total_seconds() as i32 - offset_secs);
         (MIN_INSTANT..=MAX_INSTANT)
             .contains(&ordinal_secs)
             .then_some(Self {
